+++ src/ska_mid_dish_manager/devices/DishManagerDS.py
@@ -12,7 +12,7 @@ from typing import List, Optional, Tuple
 from ska_control_model import CommunicationStatus, ResultCode
 from ska_tango_base import SKAController
 from ska_tango_base.commands import SubmittedSlowCommand
-from tango import AttrWriteType, DevState, DevULong, DevVarStringArray, DispLevel
+from tango import AttrWriteType, DevLong64, DevState, DevVarStringArray, DispLevel
 from tango.server import attribute, command, device_property, run
 
 from ska_mid_dish_manager.component_managers.dish_manager_cm import DishManagerComponentManager
@@ -1330,7 +1330,7 @@ class DishManager(SKAController):
         self.component_manager.set_noise_diode_mode(mode)
 
     @attribute(
-        dtype=(DevULong,),
+        dtype=(DevLong64,),
         max_dim_x=3,
         doc="""
             Periodic noise diode pars (units are in time quanta).
@@ -1353,7 +1353,7 @@ class DishManager(SKAController):
         self.component_manager.set_periodic_noise_diode_pars(values)
 
     @attribute(
-        dtype=(DevULong,),
+        dtype=(DevLong64,),
         max_dim_x=3,
         doc="""
             Pseudo random noise diode pars (units are in time quanta).
@@ -1598,9 +1598,7 @@ class DishManager(SKAController):
         dtype_in=bool,
         doc_in="If the synchronise argument is True, the SPFRx FPGA is instructed to synchronise "
         "its internal flywheel 1PPS to the SAT-1PPS for the ADC that is applicable to the band "
-        "being configured, and the band counters are reset. (Should be default to False). "
-        "Note when ignoring SPFRx, the configuredband on Dish.LMC will always report band B5a "
-        "when the DS indexerposition is in B5.",
+        "being configured, and the band counters are reset. (Should be default to False).",
         dtype_out=None,
         display_level=DispLevel.OPERATOR,
     )
@@ -1618,9 +1616,7 @@ class DishManager(SKAController):
         dtype_in=bool,
         doc_in="If the synchronise argument is True, the SPFRx FPGA is instructed to synchronise "
         "its internal flywheel 1PPS to the SAT-1PPS for the ADC that is applicable to the band "
-        "being configured, and the band counters are reset. (Should be default to False). "
-        "Note when ignoring SPFRx, the configuredband on Dish.LMC will always report band B5a "
-        "when the DS indexerposition is in B5.",
+        "being configured, and the band counters are reset. (Should be default to False).",
         dtype_out=None,
         display_level=DispLevel.OPERATOR,
     )