+++ CHANGELOG.rst
@@ -13,6 +13,8 @@ This project adheres to `Semantic Versioning <http://semver.org/>`_.
 - Add test_track_pattern acceptance test to run tracking test with track table input from csv files
 
   - Add manual job (k8s-test-runner-track-patterns) to gitlab pipeline to run test_track_pattern acceptance test
+- Updated periodicNoiseDiodePars and pseudoRandomNoiseDiodePars to be DevVarLong64Array - (DevLong64,)
+- Upgraded ska-mid-dish-simulators chart to v4.5.0
 
 Version 7.4.0
 *************